backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
requests
groq
psycopg2-binary